# perf(job-queue-plugin): Optimize list query for BullMQJobQueuePlugin