--- a/packages/job-queue-plugin/src/bullmq/bullmq-job-queue-strategy.ts
+++ b/packages/job-queue-plugin/src/bullmq/bullmq-job-queue-strategy.ts
@@ -31,7 +31,7 @@ import {
     loggerCtx,
     QUEUE_NAME,
 } from './constants';
-import { IndexedSetService } from './indexed-set.service';
+import { JobListIndexService } from './job-list-index.service';
 import { RedisHealthIndicator } from './redis-health-indicator';
 import { getJobsByType } from './scripts/get-jobs-by-type';
 import { BullMQPluginOptions, CustomScriptDefinition } from './types';
@@ -51,7 +51,7 @@ export class BullMQJobQueueStrategy implements InspectableJobQueueStrategy {
     private worker: Worker;
     private workerProcessor: Processor;
     private options: BullMQPluginOptions;
-    private indexedSetService: IndexedSetService;
+    private indexedSetService: JobListIndexService;
     private readonly queueNameProcessFnMap = new Map<string, (job: Job) => Promise<any>>();
     private cancellationSub: Redis;
     private readonly cancelRunningJob$ = new Subject<string>();
@@ -60,7 +60,7 @@ export class BullMQJobQueueStrategy implements InspectableJobQueueStrategy {
 
     async init(injector: Injector): Promise<void> {
         const options = injector.get<BullMQPluginOptions>(BULLMQ_PLUGIN_OPTIONS);
-        this.indexedSetService = injector.get(IndexedSetService);
+        this.indexedSetService = injector.get(JobListIndexService);
         this.options = {
             ...options,
             workerOptions: {
@@ -228,14 +228,12 @@ export class BullMQJobQueueStrategy implements InspectableJobQueueStrategy {
         let totalItems = 0;
 
         try {
-            console.time('getJobsByType');
             const [total, jobIds] = await this.callCustomScript(getJobsByType, [
                 skip,
                 take,
                 options?.filter?.queueName?.eq ?? '',
                 ...jobTypes,
             ]);
-            console.timeEnd('getJobsByType');
             items = (
                 await Promise.all(
                     jobIds.map(id => {