--- a/packages/job-queue-plugin/src/bullmq/clean-indexed-sets-task.ts
+++ b/packages/job-queue-plugin/src/bullmq/clean-indexed-sets-task.ts
@@ -1,18 +1,12 @@
 import { ScheduledTask } from '@vendure/core';
 
-import { IndexedSetService } from './indexed-set.service';
+import { JobListIndexService } from './job-list-index.service';
 
 export const cleanIndexedSetsTask = new ScheduledTask({
-    // Give your task a unique ID
-    id: 'clean-job-queue-indexed-sets',
-    // A human-readable description of the task
-    description: 'Cleans up the indexed sets used to speed up job queue listing operations',
-    // Define a default schedule. This can be modified using the
-    // `.configure()` method on the instance later.
-    schedule: cron => cron.everyDayAt(0, 0),
-    // This is the function that will be executed per the schedule.
-    async execute({injector}) {
-        const indexedSetService = injector.get(IndexedSetService);
-        return indexedSetService.cleanupIndexedSets();
+    id: 'clean-job-queue-index',
+    description: 'Cleans up the index used to speed up job queue listing operations',
+    schedule: cron => cron.everyMinute(),
+    async execute({ injector }) {
+        return injector.get(JobListIndexService).cleanupIndexedSets();
     },
 });