--- a/packages/job-queue-plugin/src/bullmq/utils.ts
+++ b/packages/job-queue-plugin/src/bullmq/utils.ts
@@ -1,0 +1,5 @@
+import { BullMQPluginOptions } from "./types";
+
+export function getPrefix(options: BullMQPluginOptions) {
+    return options.workerOptions?.prefix ?? 'bull';
+}