# Description

The current implementation for listing jobs in the BullMQJobQueuePlugin is very inefficient, which can cause resource issues on projects with hundreds of thousands of jobs.

The reason is that in order to filter on job queue name (e.g. "update-search-index"), the current Lua script loads _all_ jobs into memory from Redis, and then loops over them, using Redis `HGET` to extract the queue name from the job data.

In practice we have been seeing Redis resource problems when multiple people are sorting over a job queue with ~400k jobs, which caused overall site instability (the Redis instance was also used for session cache etc). 

This PR introduces a major performance improvement, by keeping track of jobs in dedicated sorted sets, which are ordered Redis data structures which we can use to know the contents of specific queues without reading every individual job.

This is an outline of how it works:

1. Job is created in BullMQ in the normal way
2. We then get that job ID and add it to a sorted set, named `<prefix>:queue:<queueName>:<state>` e.g. `bull:vendure-job-queue:queue:update-search-index:wait`
3. Using the BullMQ event system, we are able to move that job ID to a different sorted set when its state changes. E.g. when it gets processed and the state changes from "active" to "completed", we listen to this event and remove the ID from the `:active` sorted set and add it to the `:completed` sorted set.
4. When we want to query for the list of jobs, we only need to look at the contents of the relevant sorted set(s), which is a O(1) operation. A little extra work goes into making sure we correctly paginate and interleave sets if there are more than 1 involved (e.g. when listing all states)
5. When jobs are automatically removed from the BullMQ job queue via the `removeOnComplete` and `removeOnFail` options, we also remove the corresponding IDs from any sorted sets too. This is done via a scheduled task which by default runs every minute. On a test dataset of 250k jobs, this task takes around 1-2 seconds.

# Breaking changes

There are no breaking changes. This change should not make any difference for users, apart from the fact there is now a scheduled task that will be running every minute, doing a small amount of work (typically milliseconds on the usual amount of jobs).

Additionally, there is a method that will run on bootstrap which will automatically generate the sorted sets if they do not already exist. On the first bootstrap after upgrade, this might take a couple of seconds if there are a lot of jobs. Then it does not need to do anything again on subsequent bootstraps.

# Benchmarks

I populated my local dev server with 250,000 jobs, and then measures the time needed to generate the job list for a page of 10 with various filter criteria. Here are the results:

### 📊 getJobsByType Performance Comparison

| Run | Before (s) | After (ms) | After (s)  |
|-----|------------|------------|------------|
| 1   | 1.470      | 1.644      | 0.001644   |
| 2   | 1.463      | 4.041      | 0.004041   |
| 3   | 1.501      | 6.704      | 0.006704   |
| 4   | 1.435      | 2.775      | 0.002775   |
| 5   | 1.479      | 1.507      | 0.001507   |
| 6   | 1.454      | 1.967      | 0.001967   |
| 7   | 1.482      | 3.302      | 0.003302   |
| 8   | 1.464      | 4.030      | 0.004030   |
| 9   | 1.469      | 3.363      | 0.003363   |
| 10  | 1.502      | 1.972      | 0.001972   |
| 11  | 1.497      | 2.853      | 0.002853   |

---

### 📈 Summary

- **Average Before**: `1.475 seconds`
- **Average After**: `2.828 milliseconds` (`0.002828 seconds`)
- **Speed Increase**: `~521× faster`
- **Time Reduction**: `~99.81%`
- **Avg Time Saved per Call**: `~1.472 seconds`


# Checklist

📌 Always:
- [x] I have set a clear title
- [x] My PR is small and contains a single feature
- [x] I have [checked my own PR](## "Fix typo's and remove unused or commented out code")

👍 Most of the time:
- [ ] I have added or updated test cases
- [ ] I have updated the README if needed